requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
marshmallow==3.23.1
cryptography==43.0.3
pytest==8.3.3
pytest-flask==1.3.0
Flask-Testing==0.8.1
python-dotenv==1.0.1
Werkzeug==3.0.3